Linux: fix a gcc warning from trace print
Silence a gcc warning on 64bit arch about invalid printf
format specifier.

Warning being fixed:
--8<--8<--8<--
warning: format ‘%d’ expects argument of type ‘int’, but argument
4 has type ‘size_t {aka long unsigned int}’ [-Wformat=

+++ source/m2mconnectionhandlerpimpl.cpp
@@ -478,7 +478,8 @@ int M2MConnectionHandlerPimpl::send_to_socket(const unsigned char *buf, size_t l
         return (-1);
     }
 
-    tr_debug("send_to_socket len - %d", len);
+    // Note: "%zu" does not work on GCC+mbed-os combo
+    tr_debug("send_to_socket len - %u", (unsigned int)len);
 
     if(is_tcp_connection()) {
 #ifdef PAL_NET_TCP_AND_TLS_SUPPORT
